New "Super Quantum" cards + Bug Fixes
I've added the 4 "Super Quantum" cards released earlier this week and
fixed some bugs for "Magnaliger" and "Painful Escape". Additionally,
I've changed all prerelease "The Phantom" cards with the official
setcodes, changed all the scripts so they use the new setcode, and added
"The Phantom Knights of Shadow Veil" to the fix database with the
correct setcode.